ai_compliance/requirements.txt
# === Backend (Python) ===
flask
flask-cors
python-dotenv
openai

# === Frontend (React via Vite) ===
react
react-dom
vite
axios
